Move and document functions.
